% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub2plus.R
\name{hub2plus}
\alias{hub2plus}
\title{hub_plusを生成する}
\usage{
hub2plus(Hub_name, lato_stricto)
}
\arguments{
\item{Hub_name}{hub_name in hub_master}

\item{lato_stricto}{lato_stricto in hub_master}
}
\value{
string
}
\description{
hub_masterのhub_nameとlato_strictoから，hub_plusを生成．
「/」での区切りがある場合は，前後両方に「lato_stricto」を追加
}
\examples{
hub_master \%>\%
  tibble::as_tibble() \%>\%
  dplyr::rename_with(~stringr::str_replace_all(., "[ /]", "_")) \%>\%
  dplyr::rename_with(~stringr::str_remove_all(., "[()]")) \%>\%
  dplyr::mutate(hub_plus = hub2plus(Hub_name, lato_stricto))

}
