% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data_all}
\alias{prep_data_all}
\title{Prepare data}
\usage{
prep_data_all(path)
}
\description{
dowonload excel sheet from https://www.gbif.jp/v2/activities/wamei_checklist.html
}
\examples{
path <- "d:/wamei_checklist_ver.1.10.xlsx"
prep_data_all(path)
prep_hub_data(path)
prep_jn_data(path)
prep_ref_data()
}
