% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_jp}
\alias{ref_jp}
\title{Reference data of wamei (Japanese name) to search similar name}
\format{
A data frame with 51,809 rows and 2 variables:
\describe{
\item{source}{data souce}
\item{name_jp}{wamei (Japanese name) escaped by stringi::stri_escape_unicode()}
}
}
\usage{
ref_jp
}
\description{
Reference data of wamei (Japanese name) to search similar name
}
\keyword{datasets}
