% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_sc}
\alias{ref_sc}
\title{Reference data of scientific name to search similar name}
\format{
A data frame with 76,739 rows and 2 variables:
\describe{
\item{source}{data souce}
\item{name_sc}{scientific name}
}
}
\usage{
ref_sc
}
\description{
Reference data of scientific name to search similar name
}
\keyword{datasets}
