% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str2strvec}
\alias{str2strvec}
\title{Convert string into string vector}
\usage{
str2strvec(str, len = 1L)
}
\arguments{
\item{str}{A string to be converted.}

\item{len}{Dividing length of string.}
}
\description{
Convert string into string vector
}
