% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wamei_check.R
\name{wamei_check}
\alias{wamei_check}
\title{和名チェックリストから和名・学名の候補を出力}
\usage{
wamei_check(x, hub_master, jn_master, wide = TRUE, ds = c(GL, SF, WF, YL), ...)
}
\arguments{
\item{x}{検索する和名の文字列のベクトル．}

\item{hub_master}{和名チェックリストの「Hub_data」シートのデータ}

\item{jn_master}{和名チェックリストの「JN_dataset」シートのデータ}

\item{wide}{論理値：出力形式}

\item{ds}{データソース}
}
\value{
tibble形式．検索した和名をキーとして，和名・学名など．
維管束植物和名変換シート(excel版)に加えて，全ての和名・学名の
候補を出力．
    https://wetlands.info/tools/plantsdb/nameconv/
}
\description{
和名チェックリストを使用して，和名・学名の候補を出力します．
和名チェックリスト
    https://www.gbif.jp/v2/activities/wamei_checklist.html
}
\examples{
wamei_check(x, hub_master, jn_master, ds = c(GL, SF, WF, YL))
wamei_check(x, hub_master, jn_master, wide = FALSE, ds = WF)
}
\seealso{
wamei_check_ex()
}
