% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_hub_name.R
\name{arrange_hub_name}
\alias{arrange_hub_name}
\title{/を含むmessage・和名の整理}
\usage{
arrange_hub_name(x)
}
\arguments{
\item{x}{整理する和名}
}
\value{
string 和名の候補一覧の「，」(全角)区切り．
カタカナ部分が共通で広義と狭義ともある場合は「ワメイ広義/狭義」．
}
\description{
複数の和名(hub_plus)がある場合，重複のないように整理．
広義と狭義の両方があるときは，「/」区切りで出力．
wamei_check.R ではmessageだが，実質は和名の一覧．
}
\examples{
library(stringi)
x <- stringi::stri_unescape_unicode("\\\\u30ef\\\\u30e1\\\\u30a4/\\\\u5e83\\\\u7fa9\\\\uff0c\\\\u30ef\\\\u30e1\\\\u30a4/\\\\u72ed\\\\u7fa9")
arrange_hub_name(x)

}
