% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{editdist}
\alias{editdist}
\title{Editing distance (Levenshtein distance) of two string vectors}
\usage{
editdist(s1, s2, len = 1L)
}
\arguments{
\item{s1}{A string to be compared.}

\item{s2}{A string to be compared.}

\item{len}{Dividing length of string.}
}
\description{
Editing distance (Levenshtein distance) of two string vectors
}
