% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editdist_multi.R
\name{editdist_multi}
\alias{editdist_multi}
\alias{editdist_norm}
\title{Compute edit distance}
\usage{
editdist_multi(input, reference, inp_esc = FALSE, ref_esc = FALSE, len = 1L)

editdist_norm(s1, s2, editdist, len = 1L)
}
\arguments{
\item{input}{Vector of string to be compared.}

\item{reference}{Vector of string to be compared.}

\item{inp_esc}{Logical. TRUE when \code{input} is already escaped by
\code{stringi::stri_escape_unicode()}. Used only when \code{len} is 6.}

\item{ref_esc}{Logical. TRUE when \code{reference} is already escaped.}

\item{len}{integer 1: when checking scientific name, 6: when checking wamei (Japanese name).}

\item{s1}{A string to be compared.}

\item{s2}{A string to be compared.}

\item{editdist}{Integer. Edit distance of \code{s1} and \code{s2}.}
}
\value{
Tibble.
}
\description{
Compute edit distance
}
\section{Functions}{
\itemize{
\item \code{editdist_norm()}: Compute normalised edit distance

}}
\examples{
library(wameicheckr)
library(tidyverse)
library(stringi)
input <- 
   c("\\\\u30cf\\\\u30c3\\\\u30ab\\\\u30b0\\\\u30b5", 
     "\\\\u30b9\\\\u30ba\\\\u30ce\\\\u30a8\\\\u30f3\\\\u30c9\\\\u30a6") \%>\%
   stringi::stri_unescape_unicode()
reference <- 
   c("\\\\u30cf\\\\u30b7\\\\u30ab\\\\u30b0\\\\u30b5", 
     "\\\\u30b9\\\\u30ba\\\\u30e1\\\\u30ce\\\\u30a8\\\\u30f3\\\\u30c9\\\\u30a6") \%>\%
   stringi::stri_unescape_unicode()
editdist_multi(input = input, reference = reference, len = 6)

}
\seealso{
editdist()
}
