% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub2plus.R
\name{hub2plus}
\alias{hub2plus}
\title{hub_plusを生成する}
\usage{
hub2plus(Hub_name, lato_stricto)
}
\arguments{
\item{lato_stricto}{lato_stricto in hub_master}

\item{hub_name}{hub_name in hub_master}
}
\value{
string
}
\description{
hub_masterのhub_nameとlato_strictoから，hub_plusを生成．
「/」での区切りがある場合は，前後両方に「lato_stricto」を追加
}
\examples{
Hub_master \%>\%
 dplyr::mutate(hub_plus = hub2plus(Hub_name, lato_stricto))
}
