% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wameicheckr-package.R
\docType{package}
\name{wameicheckr-package}
\alias{wameicheckr}
\alias{wameicheckr-package}
\title{wameicheckr: Check Japanese name (wamei) of vascular plant species with R}
\description{
Check wamei of vascular plant species by wamei checklist. You can dowonload lwamei checklist from https://www.gbif.jp/v2/activities/wamei_checklist.html (Yamanouchi, T., Shutoh, K., Osawa, T., Yonekura, K., Kato, S., Shiga, T. 2019. A checklist of Japanese plant names. ).
}
\author{
\strong{Maintainer}: Toshikazu Matsumura \email{matutosi@gmail.com}

}
\keyword{internal}
