% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wameicheckr-package.R
\docType{package}
\name{wameicheckr-package}
\alias{wameicheckr}
\alias{wameicheckr-package}
\title{wameicheckr: Check Japanese name (wamei) of plant species}
\description{
Check wamei of plant species by wamei chekc list. 
    Download wamei chek list form https://www.gbif.jp/v2/activities/wamei_checklist.html
}
\author{
\strong{Maintainer}: Toshikazu Matsumura \email{matutosi@gmail.com}

}
\keyword{internal}
